fix(nrLDPC_coding_interface): Inappropriate types in synopsis of nr_(u|d)lsch_(de|en)coding_slot

+++ openair1/PHY/NR_TRANSPORT/nr_ulsch.h
@@ -79,7 +79,7 @@ int nr_ulsch_decoding_slot(PHY_VARS_gNB *phy_vars_gNB,
                            uint32_t frame,
                            uint8_t nr_tti_rx,
                            uint32_t *G,
-                           int *ULSCH_ids,
+                           uint8_t *ULSCH_ids,
                            int nb_pusch);
 
 /*! \brief Perform PUSCH unscrambling. TS 38.211 V15.4.0 subclause 6.3.1.1
